Move Category component to Mainpage folder
