+++ client/src/components/mainPage/MainPage.js
@@ -4,7 +4,7 @@ import gql from 'graphql-tag'
 import { makeStyles } from '@material-ui/styles'
 import { Card, CardContent, Typography } from '@material-ui/core'
  
-import Category from '../Category'
+import Category from './Category'
 import Loading from '../Loading'
 
 const useStyles = makeStyles({
